Lighting system revamped
Previously I used to blur the webcamTexture onto the Vtuber avatar, this was pretty expensive way of doing blurring since it rendered 5 times the vtuber and it didn't looked nice.
Now I take the webcamTexture output and paste it into a new texure with a really low size. By doing it this way we get rid of a lot of transparency issues on the vtuber avatar and we get better lighting fidelity, because this new texture size is very low we get natural blurring.
Performance wise it's not that bad honestly, by default the lighting is refreshed 8 times per second which is much better than 60 times.

Also I fixed a bug where the avatar wasn't receiving lighting if you stop and start again the lighting.